refactor(vision): rename angleUtils to angleMath, move angle difference function from poseestimator math there
