+++ src/main/java/frc/utils/math/AngleMath.java
@@ -1,9 +1,9 @@
-package frc.utils;
+package frc.utils.math;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import frc.constants.MathConstants;
 
-public class AngleUtils {
+public class AngleMath {
 
 	public static Rotation2d wrappingAbsoluteValue(Rotation2d angle) {
 		double rotations = angle.getRotations() % MathConstants.FULL_CIRCLE.getRotations();
@@ -11,6 +11,14 @@ public class AngleUtils {
 			rotations += MathConstants.FULL_CIRCLE.getRotations();
 		}
 		return Rotation2d.fromRotations(rotations);
+	}
+
+	public static Rotation2d getAngleDifference(Rotation2d angle1, Rotation2d angle2) {
+		Rotation2d difference = angle1.minus(angle2);
+		if (difference.getRadians() > Math.PI) {
+			return Rotation2d.fromRadians(MathConstants.FULL_CIRCLE.getRadians() - difference.getRadians());
+		}
+		return difference;
 	}
 
 }
